% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso.R
\name{calc_group_lasso}
\alias{calc_group_lasso}
\title{Creates a instance of group lasso model}
\usage{
calc_group_lasso(X, y, groups, result_indicator = "cp", true_betas = NULL)
}
\arguments{
\item{X}{matrix with regressors}

\item{y}{target variable}

\item{groups}{list of integers with a length equals to number of columns in X.
Indicates to which group given variable belongs to}

\item{result_indicator}{one of values ("cp", "me"). Indicates which of those two statistic 
should be used to select the final model. To use "me" also true_betas needs to be supplied.}

\item{true_betas}{array of true values of betas}
}
\value{
object of class group_lasso
}
\description{
Creates a instance of group lasso model
}
