% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_cp}
\alias{calculate_cp}
\title{Calculation of Cp value}
\usage{
calculate_cp(indexes, group_sizes, betas, betas_ls, X, y, df_function)
}
\arguments{
\item{indexes}{array with factors chosen in the model}

\item{group_sizes}{array with sizes of consecutive groups}

\item{betas}{beta coefficients in the investigated model}

\item{betas_ls}{beta coefficients in the OLS model build on the same data}

\item{X}{matrix of regressors}

\item{y}{target variable}

\item{df_function}{function that calculates degrees of freedom for specific model}
}
\value{
value of Cp statistic
}
\description{
Calculation of Cp value
}
