% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_me}
\alias{calculate_me}
\title{Calculation of model error value}
\usage{
calculate_me(X, beta_hat, beta)
}
\arguments{
\item{X}{matrix of regressors}

\item{beta_hat}{beta coefficients in the investigated model}

\item{beta}{original beta coefficients used to generate data set}
}
\value{
value of model error
}
\description{
Calculation of model error value
}
