% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{calculate_test}
\alias{calculate_test}
\title{Function used to test specific method}
\usage{
calculate_test(name, test_function, n, create_model, ...)
}
\arguments{
\item{name}{string name which will be visible in the results}

\item{test_function}{function with (X, y, true_betas, groups) as input and returning 
test_result object}

\item{n}{how many times the test will be run}

\item{create_model}{function that returns data set (X, y, groups)}

\item{...}{additional arguments for test_function}
}
\value{
test_results object
}
\description{
Function used to test specific method
}
