% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{categorize_matrix}
\alias{categorize_matrix}
\title{Trichotomization of values in the matrix}
\usage{
categorize_matrix(Z)
}
\arguments{
\item{Z}{matrix}
}
\value{
matrix with trichomizated values
}
\description{
Trichotomization of values in the matrix
}
