% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{count_factors}
\alias{count_factors}
\title{Function used to calculate unique factors}
\usage{
count_factors(betas, betas_names)
}
\arguments{
\item{betas}{values of coefficients}

\item{betas_names}{names of the coefficients}
}
\value{
number of unique factors
}
\description{
Function used to calculate unique factors
}
