% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_boxplot,test_container,character-method}
\alias{create_boxplot,test_container,character-method}
\title{Creates boxplot for test_container.}
\usage{
\S4method{create_boxplot}{test_container,character}(container, column)
}
\arguments{
\item{container}{test_container. Instance of test_container_class}

\item{column}{character. Column which values will be presented in the boxplot. 
One of the ("model_error", "n_factors", "cpu_time")}
}
\value{
ggplot2 object with boxplot
}
\description{
Creates boxplot for test_container.
}
