% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_model1}
\alias{create_model1}
\title{Creation of type 1 data set}
\usage{
create_model1(n = 50, p = 15)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}
}
\value{
list with three elements - X: design matrix, y: target variable,
betas: coefficients used to create y
}
\description{
Creation of type 1 data set
}
