% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_model3}
\alias{create_model3}
\title{Creation of type 3 data set}
\usage{
create_model3(n = 100, p = 16)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}
}
\value{
list with three elements - X: design matrix, y: target variable,
betas: coefficients used to create y
}
\description{
Creation of type 3 data set
}
