% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_model4}
\alias{create_model4}
\title{Creation of type 4 data set}
\usage{
create_model4(n = 100, p1 = 10, p2 = 10)
}
\arguments{
\item{n}{number of observations}

\item{p1}{number of discrete variables}
}
\value{
list with three elements - X: design matrix, y: target variable,
betas: coefficients used to create y
}
\description{
Creation of type 4 data set
}
