% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_table,test_container-method}
\alias{create_table,test_container-method}
\title{Creates table with aggregated results of the tests}
\usage{
\S4method{create_table}{test_container}(container)
}
\arguments{
\item{container}{test_container. Instance of class test_container}
}
\value{
data frame with results of all tests. This table's shape is based on the results in the
article.
}
\description{
Creates table with aggregated results of the tests
}
