% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lars.R
\name{df_lars}
\alias{df_lars}
\title{Calculates degrees of freedom for group lars model.}
\usage{
df_lars(indexes, group_sizes, betas, betas_ls)
}
\arguments{
\item{indexes}{array with factors chosen in the model}

\item{group_sizes}{array with sizes of consecutive groups}

\item{betas}{beta coefficients in the investigated model}

\item{betas_ls}{beta coefficients in the OLS model build on the same data}
}
\value{
number of degrees of freedom
}
\description{
Calculates degrees of freedom for group lars model.
}
