% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lars.R
\name{find_alpha_lars}
\alias{find_alpha_lars}
\title{Finds optimum for quadratic equation needed to find next factor included in the LARS algorithm.}
\usage{
find_alpha_lars(X, r, j, mcs, gamma_)
}
\arguments{
\item{X}{matrix with regressors}

\item{r}{current residuals}

\item{j}{candidate factor}

\item{mcs}{current "active set"}

\item{gamma_}{current direction}
}
\value{
value of root which is in [0,1] interval
}
\description{
Finds optimum for quadratic equation needed to find next factor included in the LARS algorithm.
}
