% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{first_up}
\alias{first_up}
\title{Makes first letter of string uppercase}
\usage{
first_up(x)
}
\arguments{
\item{x}{string}
}
\value{
transformed string
}
\description{
Makes first letter of string uppercase
}
