% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{generate_noise}
\alias{generate_noise}
\title{Generating noise for target variable}
\usage{
generate_noise(Y, ratio)
}
\arguments{
\item{Y}{array}

\item{ratio}{signal-to-noise-ratio}
}
\value{
array with noise
}
\description{
Generating noise for target variable
}
