% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_test,test_container,character-method}
\alias{get_test,test_container,character-method}
\title{Test getter}
\usage{
\S4method{get_test}{test_container,character}(container, name)
}
\arguments{
\item{container}{test_container. Instance of test_container class}

\item{name}{character. Name of the test to be returned}
}
\value{
instance of class test_results from the container. If there is no test with such a name
method will throw an error.
}
\description{
Test getter
}
