% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%-\%}
\alias{\%-\%}
\title{A easier form of setting part of array to zero}
\usage{
vector \%-\% index
}
\arguments{
\item{vector}{array which is meant to be used}

\item{index}{indexes were zeros will be inserted}
}
\value{
array with zeros in selected indexes
}
\description{
Implementation of b_{-j}=(b^'_1, ..., b^'_{j-1}, 0', b^'_{j+1}, ..., b^'_{J})
}
