% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{norm_L}
\alias{norm_L}
\title{Vector norm mentioned in the article}
\usage{
norm_L(vector, p)
}
\arguments{
\item{vector}{array}

\item{p}{multiplier of identity matrix}
}
\value{
vector norm
}
\description{
Vector norm mentioned in the article
}
