% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{perform_ttest,test_container,character,character-method}
\alias{perform_ttest,test_container,character,character-method}
\title{Performs check if results of the models are statistically different}
\usage{
\S4method{perform_ttest}{test_container,character,character}(container, tests_rows, tests_cols)
}
\arguments{
\item{container}{test_container. Instance of test_container class}

\item{tests_rows}{character. One group of tests (may be an array). Will be presented in the rows}

\item{tests_cols}{character. Second group of tests (may be an array). 
Will be presented in the columns}
}
\value{
table with p-values of t-test.
}
\description{
Performs check if results of the models are statistically different
}
