% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{+,test_container,test_results-method}
\alias{+,test_container,test_results-method}
\title{Adding test_results to container}
\usage{
\S4method{+}{test_container,test_results}(e1, e2)
}
\arguments{
\item{e1}{test_container. Instance of class test_container}

\item{e2}{test_results. Instance of class test_results}
}
\value{
instance of class test_container with added new test_results.
}
\description{
Adding test_results to container
}
