% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lars.R
\name{quad_roots}
\alias{quad_roots}
\title{Very simple quadratic equation solver}
\usage{
quad_roots(a, b, c)
}
\arguments{
\item{a}{quadratic coefficient}

\item{b}{linear coefficient}

\item{c}{constant coefficient}
}
\value{
array with two roots
}
\description{
Very simple quadratic equation solver
}
