% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{test_container-class}
\alias{test_container-class}
\title{Object that stores instances of tests_results}
\value{
object of class test_container
}
\description{
Object that stores instances of tests_results
}
\section{Slots}{

\describe{
\item{\code{tests}}{list.}
}}

