% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_lars_group}
\alias{test_lars_group}
\title{Function that test group lars method}
\usage{
test_lars_group(X, y, true_betas, groups, ...)
}
\arguments{
\item{X}{design matrix}

\item{y}{target variable}

\item{true_betas}{beta coefficients used in Y calculation}

\item{groups}{list of integers with a length equals to number of columns in X.
Indicates to which group given variable belongs to}
}
\value{
instance of test_result class
}
\description{
Function that test group lars method
}
