% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_ls}
\alias{test_ls}
\title{Function that test OLS method}
\usage{
test_ls(X, y, true_betas, groups)
}
\arguments{
\item{X}{design matrix}

\item{y}{target variable}

\item{true_betas}{beta coefficients used in Y calculation}

\item{groups}{added only to keep function's shape}
}
\value{
instance of test_result class
}
\description{
Function that test OLS method
}
