% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{test_result-class}
\alias{test_result-class}
\title{Title}
\value{
instance of test_result class
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{model_error}}{numeric. Model error obtained in the test}

\item{\code{n_factors}}{numeric. Number of factors obtained in the test}

\item{\code{cpu_time}}{numeric. CPU time obtained in the test}
}}

