% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{test_results-class}
\alias{test_results-class}
\title{Class containing information from multiple tests runs}
\value{
instance of test_results class
}
\description{
Class containing information from multiple tests runs
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the model}

\item{\code{model_error}}{numeric. Mean model error}

\item{\code{model_error_list}}{numeric. All model errors obtained during testing}

\item{\code{model_error_std}}{numeric. Standard deviation of model error}

\item{\code{n_factors}}{numeric. Mean number of factors}

\item{\code{n_factors_list}}{integer. All numbers of factors obtained during testing}

\item{\code{n_factors_std}}{numeric. Standard deviation of number of factors}

\item{\code{cpu_time}}{numeric. Mean CPU time}

\item{\code{cpu_time_list}}{numeric. All CPU times obtained during testing}

\item{\code{cpu_time_std}}{numeric. Standard deviation of CPU time}
}}

