% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{generate_noise}
\alias{generate_noise}
\title{generate_noise function}
\usage{
generate_noise(Y, ratio)
}
\arguments{
\item{Y}{a}

\item{ratio}{b}
}
\value{
something
}
\description{
generate_noise function
}
